requirements.txt
image==1.5.20
numpy==1.14.3
tensorflow==1.14.0
Pillow==6.0.0
